complete rework of the checkPropertiesFun-API
A checkPropertiesFun now accepts the properties to check directly as
arguments and not merely as a single list of key-value pairs. This
allows for positional and partial matching as well as for cleaner code.

Furthermore, a checkPropertiesFun is now allowed to modify the
properties, e.g. to make them consistent, which in turn can speed up the
actual type checking.