custom types can now be registered per package
Custom types are now stored in the parent frame of
registerCustomType/Alias.

When lazyTyper searches for custom types/alias during declare, cast
and registerCustomType/Alias, it searches the enclosing environment(s)
of the calling function (if any) until it finds the type or until it
encounters a namespace (or the global environment).